Force le 32bits sur la gestion des adresse IP
git-svn-id: svn://localhost/qrezix/branches/qrezix--qt4/qrezix@826 30024bfc-89fa-0310-95ff-ee513f6fc940

+++ rzxconfig.cpp
@@ -767,7 +767,7 @@ const QPixmap &RzxConfig::localhostIcon(){
 	return icon("localhost");
 }
 
-QString RzxConfig::historique(unsigned long ip, const QString& hostname) {
+QString RzxConfig::historique(quint32 ip, const QString& hostname) {
 	QString filename = QString::number(ip, 16);
 	
 	QDir logdir = globalConfig() -> logDir();
